requirements.txt
streamlit==1.32.0
anthropic==0.49.0
python-dotenv==1.0.1
loguru==0.7.2